Add TracebackFilter and NoOpTracebackFilter for improved traceback handling
