+++ vedro/core/exc_info/__init__.py
@@ -1,3 +1,4 @@
 from ._exc_info import ExcInfo
+from ._traceback_filter import NoOpTracebackFilter, TracebackFilter, TracebackFilterType
 
-__all__ = ("ExcInfo",)
+__all__ = ("ExcInfo", "TracebackFilter", "NoOpTracebackFilter", "TracebackFilterType",)
